+++ realip.go
@@ -1,26 +1,10 @@
 package rest
 
 import (
-	"bytes"
-	"fmt"
-	"net"
 	"net/http"
-	"strings"
+
+	"github.com/go-pkgz/rest/utils"
 )
-
-type ipRange struct {
-	start net.IP
-	end   net.IP
-}
-
-var privateRanges = []ipRange{
-	{start: net.ParseIP("10.0.0.0"), end: net.ParseIP("10.255.255.255")},
-	{start: net.ParseIP("100.64.0.0"), end: net.ParseIP("100.127.255.255")},
-	{start: net.ParseIP("172.16.0.0"), end: net.ParseIP("172.31.255.255")},
-	{start: net.ParseIP("192.0.0.0"), end: net.ParseIP("192.0.0.255")},
-	{start: net.ParseIP("192.168.0.0"), end: net.ParseIP("192.168.255.255")},
-	{start: net.ParseIP("198.18.0.0"), end: net.ParseIP("198.19.255.255")},
-}
 
 // RealIP is a middleware that sets a http.Request's RemoteAddr to the results
 // of parsing either the X-Forwarded-For or X-Real-IP headers.
@@ -31,7 +15,7 @@ var privateRanges = []ipRange{
 // as X-Forwarded-For header and attack the server in various ways.
 func RealIP(h http.Handler) http.Handler {
 	fn := func(w http.ResponseWriter, r *http.Request) {
-		if rip, err := GetIPAddress(r); err == nil {
+		if rip, err := utils.GetIPAddress(r); err == nil {
 			r.RemoteAddr = rip
 		}
 		h.ServeHTTP(w, r)
@@ -40,58 +24,10 @@ func RealIP(h http.Handler) http.Handler {
 	return http.HandlerFunc(fn)
 }
 
-// GetIPAddress returns real ip from the given request
-func GetIPAddress(r *http.Request) (string, error) {
-
-	for _, h := range []string{"X-Forwarded-For", "X-Real-Ip"} {
-		addresses := strings.Split(r.Header.Get(h), ",")
-		// march from right to left until we get a public address
-		// that will be the address right before our proxy.
-		for i := len(addresses) - 1; i >= 0; i-- {
-			ip := strings.TrimSpace(addresses[i])
-			realIP := net.ParseIP(ip)
-			if !realIP.IsGlobalUnicast() || isPrivateSubnet(realIP) {
-				continue
-			}
-			return ip, nil
-		}
+// GetIPAddress returns real ip from the given request, if ip can be extracted returns ""
+func GetIPAddress(r *http.Request) string {
+	if rip, err := utils.GetIPAddress(r); err == nil {
+		return rip
 	}
-
-	// X-Forwarded-For header set but parsing failed above
-	if r.Header.Get("X-Forwarded-For") != "" {
-		return "", fmt.Errorf("no valid ip found")
-	}
-
-	// get IP from RemoteAddr
-	ip, _, err := net.SplitHostPort(r.RemoteAddr)
-	if err != nil {
-		return "", fmt.Errorf("can't parse ip %q: %w", r.RemoteAddr, err)
-	}
-	if netIP := net.ParseIP(ip); netIP == nil {
-		return "", fmt.Errorf("no valid ip found")
-	}
-
-	return ip, nil
-}
-
-// inRange - check to see if a given ip address is within a range given
-func inRange(r ipRange, ipAddress net.IP) bool {
-	// strcmp type byte comparison
-	if bytes.Compare(ipAddress, r.start) >= 0 && bytes.Compare(ipAddress, r.end) < 0 {
-		return true
-	}
-	return false
-}
-
-// isPrivateSubnet - check to see if this ip is in a private subnet
-func isPrivateSubnet(ipAddress net.IP) bool {
-	if ipCheck := ipAddress.To4(); ipCheck != nil {
-		for _, r := range privateRanges {
-			// check if this ip is in a private range
-			if inRange(r, ipAddress) {
-				return true
-			}
-		}
-	}
-	return false
+	return ""
 }
