+++ realip_test.go
@@ -17,8 +17,7 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Real-IP", "8.8.8.8")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "8.8.8.8", adr)
 	}
 	{
@@ -26,8 +25,7 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "8.8.8.8,1.1.1.2, 30.30.30.1")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -35,8 +33,7 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "8.8.8.8,1.1.1.2,192.168.1.1,10.0.0.65")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "1.1.1.2", adr)
 	}
 	{
@@ -45,8 +42,7 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "30.30.30.1")
 		req.Header.Add("X-Real-Ip", "10.0.0.1")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -55,8 +51,7 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "30.30.30.1")
 		req.Header.Add("X-Real-Ip", "8.8.8.8")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -65,23 +60,21 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "10.0.0.2,192.168.1.1")
 		req.Header.Add("X-Real-Ip", "8.8.8.8")
-		adr, err := GetIPAddress(req)
-		require.NoError(t, err)
+		adr := GetIPAddress(req)
 		assert.Equal(t, "8.8.8.8", adr)
 	}
 	{
 		req, err := http.NewRequest("GET", "/something", http.NoBody)
 		assert.NoError(t, err)
-		_, err = GetIPAddress(req)
-		require.Error(t, err)
+		ip := GetIPAddress(req)
+		assert.Equal(t, "", ip)
 	}
 }
 
 func TestGetIPAddressFromRemoteAddr(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		log.Printf("%v", r)
-		adr, err := GetIPAddress(r)
-		require.NoError(t, err)
+		adr := GetIPAddress(r)
 		assert.Equal(t, "127.0.0.1", adr)
 	}))
 
@@ -97,8 +90,7 @@ func TestRealIP(t *testing.T) {
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		log.Printf("%v", r)
 		require.Equal(t, "1.2.3.4", r.RemoteAddr)
-		adr, err := GetIPAddress(r)
-		require.NoError(t, err)
+		adr := GetIPAddress(r)
 		assert.Equal(t, "1.2.3.4", adr)
 	})
